Aether

Description

Aether provides a high-level interface for fetching information from Wikipedia, a free online encyclopedia. Aether can be used to download articles, get search results, generate random article titles, and more.

Aether wraps around the MediaWiki API. Learn more at http://www.mediawiki.org/wiki/API.

Documentation

licenses :: String

Returns information regarding Wikipedia's text license (CC BY-SA) and the license used by Aether (MIT).

search :: String -> IO [String]

Returns a list of up to 10 Wikipedia search results based on the given search terms.

suggest :: String -> IO String

Returns a suggestion based on the given search terms. Returns an empty string if no suggestion is available.

random :: Int -> IO [String]

Returns a list of random Wikipedia article titles. The given Int determines the length of the list; up to 10 titles can be fetched at once. The returned list is guaranteed to contain no duplicate titles.

summary :: String -> IO String

Returns a summary of the Wikipedia article with the given title. The summary will be about one line long.

summaryLines :: Int -> String -> IO String

Returns a summary of the Wikipedia article with the given title. The given Int determines the maximum summary length, in lines.

page :: String -> IO WikipediaPage

Returns a WikipediaPage for the article with the given title. Returns a WikipediaPage with all fields null in the event of an error. For error handling using Maybe instead, use pageMaybe.

pageMaybe :: String -> IO (Maybe WikipediaPage)

Returns a WikipediaPage for the article with the given title. Errors are handled using Maybe (Nothing is returned on error).