Aether.Parser

Description

This module contains several simple XML-parsing regex functions, as well as other text utility functions.

Documentation

extractBetween :: String -> String -> String -> String

extractBetween text start end will return the portion of text between start and end as a string.

extractAllAttrValues :: String -> String -> [String]

extractAllAttrValues text attr, where text is a XML string, returns the values of all instances of the XML attribute attr.

extractAttrValue :: String -> String -> String

extractAttrValue text attr, where text is a XML string, returns the value of the first instance of the XML attribute attr. Returns an empty string If no instances of attr exist in the given XML string.

trim :: String -> String

Returns the given string with whitespace trimmed.