Aether.WebService

Description

This module contains the networking functions used to access Wikipedia via the MediaWiki API. More information on the MediaWiki API can be found at http://www.mediawiki.org/wiki/API.

Note that arguments for API queries are represented as key-value pairs of type [(String, String)].

Documentation

stdQueries :: [(String, String)]

Common query arguments that should be included in most API queries. Defined as stdQueries = [("format", "xml"), ("action", "query")]. Consult the MediaWiki API for query parameter details.

queriesToURI :: [(String, String)] -> URI

Returns an API query URI based on the given query arguments.

wikiRequest :: [(String, String)] -> IO (Maybe String)

Sends an API query using the given query arguments and returns the result as a string wrapped by Maybe. (Nothing is returned in the event of a HTTP request error.)