Aether.WikipediaPage

Description

This module exports the WikipediaPage data type for representing Wikipedia pages and functions related to it.

Documentation

images :: WikipediaPage -> IO (Maybe [String])

Returns the URLs of the images on the given page. Works for up to 500 images.

imageDescs :: WikipediaPage -> IO (Maybe [String])

Returns the URLs of the description pages for the images on the given page. Works for up to 500 images.

isRedirect :: WikipediaPage -> Bool

Returns if the given page is a hard redirect.

sections :: WikipediaPage -> IO (Maybe [String])

Returns the section titles of the given page.

data WikipediaPage

Represents a single Wikipedia page.

Constructors

WikipediaPage

Fields

title :: String

Page title (with namespace prefix).

content :: String

Raw contents (wiki markup).

pageID :: String

Page ID.

lastEdit :: String

Timestamp of last edit.

queryURI :: String

URI of the API query used.